Add files via upload
Here is the code for my final project. I originally had it so the left sensor would trigger the left ear and the right sensor would trigger the right ear, but I commented out the latter part of the code and just made it so either sensor would trigger both. I did this because I wanted it to work so if you stroked both sensors at the same time they would both go off, but it didn't work that way so I changed it.